Reorganize mock resolvers
